Remove categories app
