--- orb/categories/tests/factory.py
@@ -1,9 +1,0 @@
-from orb.models import Category
-
-
-def category_factory(**kwargs):
-    defaults = {
-        "name": "Test category",
-    }
-    defaults.update(kwargs)
-    return Category.objects.create(**defaults)
